examples tidying: remove kokkostools example and hardcoded values from example inputfiles and add toggles for specific parts of inputfiles

+++ examples/boxmodelcollisions/breakup.py
@@ -116,6 +116,9 @@ def inputfiles(
     config_filename,
     config_params,
     kernel,
+    gen_config,
+    gen_gbxs,
+    gen_supers,
     isfigures,
 ):
     from cleopy import editconfigfile
@@ -131,16 +134,19 @@ def inputfiles(
         savefigpath.mkdir(exist_ok=True)
 
     ### --- copy src_config_filename into tmp and edit parameters --- ###
-    config_filename.unlink(missing_ok=True)  # delete any existing config
-    shutil.copy(src_config_filename, config_filename)
-    editconfigfile.edit_config_params(config_filename, config_params)
+    if gen_config:
+        config_filename.unlink(missing_ok=True)  # delete any existing config
+        shutil.copy(src_config_filename, config_filename)
+        editconfigfile.edit_config_params(config_filename, config_params)
 
     ### --- delete any existing initial conditions --- ###
     yaml = YAML()
     with open(config_filename, "r") as file:
         config = yaml.load(file)
-    Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
-    Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
+    if gen_gbxs:
+        Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
+    if gen_supers:
+        Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
 
     ### --- input binary files generation --- ###
     # equivalent to ``import breakup_inputfiles`` followed by
@@ -157,6 +163,10 @@ def inputfiles(
         config_filename,
         kernel,
     ]
+    if gen_gbxs:
+        cmd.append("--gen_gbxs")
+    if gen_supers:
+        cmd.append("--gen_supers")
     if isfigures[0]:
         cmd.append("--show_figures")
     if isfigures[1]:
@@ -217,6 +227,9 @@ def plot_results(path2CLEO, savefigpath, kernels, config_filenames):
 ### --------------------- RUN EXAMPLE FOR EACH KERNEL ---------------------- ###
 for kernel, [config_filename, config_params] in kernel_configs.items():
     if args.do_inputfiles:
+        gen_config = True
+        gen_gbxs = True
+        gen_supers = True
         inputfiles(
             path2CLEO,
             path2build,
@@ -228,6 +241,9 @@ for kernel, [config_filename, config_params] in kernel_configs.items():
             config_filename,
             config_params,
             kernel,
+            gen_config,
+            gen_gbxs,
+            gen_supers,
             isfigures,
         )
